Move CfgWeapons to dedicated .hpp file
